updating openLoginModal redux event to open new login modal
