+++ components/UnifiedDocFeed/FeedBlurWithButton.tsx
@@ -71,11 +71,10 @@ const styles = StyleSheet.create({
     cursor: "pointer",
     boxSizing: "border-box",
     width: "unset",
-    padding: "0px 15px",
+    // padding: "0px 15px",
     boxShadow: "0 0 15px rgba(0, 0, 0, 0.14)",
     "@media only screen and (max-width: 415px)": {
       height: 50,
-      width: 140,
       fontSize: 18,
     },
   },
