+++ pages/org/join/[token]/index.js
@@ -8,10 +8,10 @@ import Loader from "~/components/Loader/Loader";
 import { MessageActions } from "~/redux/message";
 import colors from "~/config/themes/colors";
 import OrgAvatar from "~/components/Org/OrgAvatar";
-import GoogleLoginButton from "~/components/GoogleLoginButton";
 import { AuthActions } from "~/redux/auth";
 import HeadComponent from "~/components/Head";
 import { captureEvent } from "~/config/utils/events";
+import Login from "~/components/Login/Login";
 
 const Index = ({ auth, showMessage, setMessage, googleLogin, getUser }) => {
   const router = useRouter();
@@ -86,13 +86,11 @@ const Index = ({ auth, showMessage, setMessage, googleLogin, getUser }) => {
               hideRipples={true}
             />
           ) : (
-            <GoogleLoginButton
-              styles={styles.googleLoginButton}
-              googleLogin={googleLogin}
-              getUser={getUser}
-              loginCallback={handleJoinOrg}
-              customLabel={"Sign in with Google to join"}
-            />
+            <Login loginCallback={handleJoinOrg}>
+              <Button hideRipples={true} fullWidth size="large">
+                <span>Sign in to join</span>
+              </Button>
+            </Login>
           )}
         </div>
       )}
